Check if SDL_main is in use
The main() function of colobot-app is declared as extern "C" to prevent
name mangling. This is required because on some platforms, SDL2 declares
its own main() function and defines a macro that renames the user's main
to SDL_main; in which case, name mangling may cause linking failures.

However, when building for platforms where this is not the case,
gcc15 complains that specifying linkage for main is not allowed.
> error: cannot declare ‘::main’ with a linkage
>        specification [-Wpedantic]

This commit wraps the extern block in #ifdefs that check
if the main -> SDL_main macro is in use.